backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx>=0.26.0,<0.28
python-dotenv>=1.1.0
supabase==2.9.1
pydantic[email]>=2.11.7,<3.0
pydantic-settings>=2.6.1,<3.0
requests==2.31.0
solana==0.36.6
anchorpy==0.21.0
anchorpy-idl==0.1.0

